+++ src/gst/gstpipewiresink.c
@@ -533,7 +533,7 @@ gst_pipewire_sink_setcaps (GstBaseSink * bsink, GstCaps * caps)
 
     pw_stream_connect (pwsink->stream,
                           PW_DIRECTION_OUTPUT,
-                          pwsink->path ? atoi(pwsink->path) : SPA_ID_INVALID,
+                          pwsink->path ? (uint32_t)atoi(pwsink->path) : SPA_ID_INVALID,
                           flags,
                           (const struct spa_pod **) possible->pdata,
                           possible->len);
